Automatic brightness adjustment based on sunset/sunrise
Lower the brightness of the LEDs during the night.  Night
is based as the time between sunset and sunrise and the
fade in/fade out follows a guassian function over the time.

There is also a artificially defined night time during
which time the leds are dimmed even more.  The fade in/
fade out during this time follows a bezier curve with
a rapid fade out and a slow fade in.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>